app/functionapp/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
# azurefunctions-extensions-bindings-blob
azure-storage-blob
openai
aiohttp
azure-identity
azure-cosmos
token-count==0.2.1
